# swagger/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api/ocp-requirement-api/ocp-requirement-api.proto",
    "version": "version not set"
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/requirements": {
      "get": {
        "operationId": "OcpRequirementApi_ListRequirementsV1",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiListRequirementsV1Response"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "uint64"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "uint64"
          }
        ],
        "tags": [
          "OcpRequirementApi"
        ]
      },
      "post": {
        "operationId": "OcpRequirementApi_CreateRequirementV1",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiCreateRequirementV1Response"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiCreateRequirementV1Request"
            }
          }
        ],
        "tags": [
          "OcpRequirementApi"
        ]
      }
    },
    "/v1/requirements/{requirement_id}": {
      "get": {
        "operationId": "OcpRequirementApi_DescribeRequirementV1",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiDescribeRequirementV1Response"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "requirement_id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uint64"
          }
        ],
        "tags": [
          "OcpRequirementApi"
        ]
      },
      "delete": {
        "operationId": "OcpRequirementApi_RemoveRequirementV1",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiRemoveRequirementV1Response"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "requirement_id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uint64"
          }
        ],
        "tags": [
          "OcpRequirementApi"
        ]
      }
    }
  },
  "definitions": {
    "apiCreateRequirementV1Request": {
      "type": "object",
      "properties": {
        "user_id": {
          "type": "string",
          "format": "uint64"
        },
        "text": {
          "type": "string"
        }
      }
    },
    "apiCreateRequirementV1Response": {
      "type": "object",
      "properties": {
        "requirement_id": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
    "apiDescribeRequirementV1Response": {
      "type": "object",
      "properties": {
        "requirement": {
          "$ref": "#/definitions/apiRequirement"
        }
      }
    },
    "apiListRequirementsV1Response": {
      "type": "object",
      "properties": {
        "requirements": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apiRequirement"
          }
        }
      }
    },
    "apiRemoveRequirementV1Response": {
      "type": "object",
      "properties": {
        "found": {
          "type": "boolean"
        }
      }
    },
    "apiRequirement": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uint64"
        },
        "user_id": {
          "type": "string",
          "format": "uint64"
        },
        "text": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "type_url": {
          "type": "string"
        },
        "value": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "runtimeError": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string"
        },
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}